+++ frontend/src/components/GddTransactionListFooter.spec.js
