+++ frontend/src/components/GddTransactionListFooter.vue
